+++ pkg/commands/alias/serviceversion/doc.go
@@ -1,0 +1,2 @@
+// Package serviceversion contains the 'service-version' aliases for the 'service version' commands.
+package serviceversion