+++ pkg/commands/service/version/clone.go
@@ -1,8 +1,7 @@
-package serviceversion
+package version
 
 import (
 	"context"
-	"fmt"
 	"io"
 
 	"github.com/fastly/go-fastly/v12/fastly"
@@ -13,25 +12,22 @@ import (
 	"github.com/fastly/cli/pkg/text"
 )
 
-// UpdateCommand calls the Fastly API to update a service version.
-type UpdateCommand struct {
+// CloneCommand calls the Fastly API to clone a service version.
+type CloneCommand struct {
 	argparser.Base
-	input          fastly.UpdateVersionInput
+	argparser.JSONOutput
+
+	Input          fastly.CloneVersionInput
 	serviceName    argparser.OptionalServiceNameID
 	serviceVersion argparser.OptionalServiceVersion
-	autoClone      argparser.OptionalAutoClone
-
-	comment argparser.OptionalString
 }
 
-// NewUpdateCommand returns a usable command registered under the parent.
-func NewUpdateCommand(parent argparser.Registerer, g *global.Data) *UpdateCommand {
-	c := UpdateCommand{
-		Base: argparser.Base{
-			Globals: g,
-		},
-	}
-	c.CmdClause = parent.Command("update", "Update a Fastly service version")
+// NewCloneCommand returns a usable command registered under the parent.
+func NewCloneCommand(parent argparser.Registerer, g *global.Data) *CloneCommand {
+	var c CloneCommand
+	c.Globals = g
+	c.CmdClause = parent.Command("clone", "Clone a Fastly service version")
+	c.RegisterFlagBool(c.JSONFlag()) // --json
 	c.RegisterFlag(argparser.StringFlagOpts{
 		Name:        argparser.FlagServiceIDName,
 		Description: argparser.FlagServiceIDDesc,
@@ -50,24 +46,15 @@ func NewUpdateCommand(parent argparser.Registerer, g *global.Data) *UpdateComman
 		Dst:         &c.serviceVersion.Value,
 		Required:    true,
 	})
-	c.RegisterAutoCloneFlag(argparser.AutoCloneFlagOpts{
-		Action: c.autoClone.Set,
-		Dst:    &c.autoClone.Value,
-	})
-
-	// TODO(integralist):
-	// Make 'comment' field mandatory once we roll out a new release of Go-Fastly
-	// which will hopefully have better/more correct consistency as far as which
-	// fields are supposed to be optional and which should be 'required'.
-	//
-	c.CmdClause.Flag("comment", "Human-readable comment").Action(c.comment.Set).StringVar(&c.comment.Value)
 	return &c
 }
 
 // Exec invokes the application logic for the command.
-func (c *UpdateCommand) Exec(_ io.Reader, out io.Writer) error {
+func (c *CloneCommand) Exec(_ io.Reader, out io.Writer) error {
+	if c.Globals.Verbose() && c.JSONOutput.Enabled {
+		return errors.ErrInvalidVerboseJSONCombo
+	}
 	serviceID, serviceVersion, err := argparser.ServiceDetails(argparser.ServiceDetailsOpts{
-		AutoCloneFlag:      c.autoClone,
 		APIClient:          c.Globals.APIClient,
 		Manifest:           *c.Globals.Manifest,
 		Out:                out,
@@ -83,23 +70,21 @@ func (c *UpdateCommand) Exec(_ io.Reader, out io.Writer) error {
 		return err
 	}
 
-	c.input.ServiceID = serviceID
-	c.input.ServiceVersion = fastly.ToValue(serviceVersion.Number)
-	if !c.comment.WasSet {
-		return fmt.Errorf("error parsing arguments: required flag --comment not provided")
-	}
-	c.input.Comment = &c.comment.Value
+	c.Input.ServiceID = serviceID
+	c.Input.ServiceVersion = fastly.ToValue(serviceVersion.Number)
 
-	ver, err := c.Globals.APIClient.UpdateVersion(context.TODO(), &c.input)
+	ver, err := c.Globals.APIClient.CloneVersion(context.TODO(), &c.Input)
 	if err != nil {
 		c.Globals.ErrLog.AddWithContext(err, map[string]any{
 			"Service ID":      serviceID,
 			"Service Version": fastly.ToValue(serviceVersion.Number),
-			"Comment":         c.comment.Value,
 		})
 		return err
 	}
 
-	text.Success(out, "Updated service %s version %d", fastly.ToValue(ver.ServiceID), c.input.ServiceVersion)
+	if ok, err := c.WriteJSON(out, ver); ok {
+		return err
+	}
+	text.Success(out, "Cloned service %s version %d to version %d", fastly.ToValue(ver.ServiceID), c.Input.ServiceVersion, fastly.ToValue(ver.Number))
 	return nil
 }