+++ pkg/commands/service/version/lock.go
@@ -1,4 +1,4 @@
-package serviceversion
+package version
 
 import (
 	"context"
@@ -14,20 +14,19 @@ import (
 	"github.com/fastly/cli/pkg/text"
 )
 
-// ActivateCommand calls the Fastly API to activate a service version.
-type ActivateCommand struct {
+// LockCommand calls the Fastly API to lock a service version.
+type LockCommand struct {
 	argparser.Base
-	Input          fastly.ActivateVersionInput
+	Input          fastly.LockVersionInput
 	serviceName    argparser.OptionalServiceNameID
 	serviceVersion argparser.OptionalServiceVersion
-	autoClone      argparser.OptionalAutoClone
 }
 
-// NewActivateCommand returns a usable command registered under the parent.
-func NewActivateCommand(parent argparser.Registerer, g *global.Data) *ActivateCommand {
-	var c ActivateCommand
+// NewLockCommand returns a usable command registered under the parent.
+func NewLockCommand(parent argparser.Registerer, g *global.Data) *LockCommand {
+	var c LockCommand
 	c.Globals = g
-	c.CmdClause = parent.Command("activate", "Activate a Fastly service version")
+	c.CmdClause = parent.Command("lock", "Lock a Fastly service version")
 	c.RegisterFlag(argparser.StringFlagOpts{
 		Name:        argparser.FlagServiceIDName,
 		Description: argparser.FlagServiceIDDesc,
@@ -46,18 +45,13 @@ func NewActivateCommand(parent argparser.Registerer, g *global.Data) *ActivateCo
 		Dst:         &c.serviceVersion.Value,
 		Required:    true,
 	})
-	c.RegisterAutoCloneFlag(argparser.AutoCloneFlagOpts{
-		Action: c.autoClone.Set,
-		Dst:    &c.autoClone.Value,
-	})
 	return &c
 }
 
 // Exec invokes the application logic for the command.
-func (c *ActivateCommand) Exec(_ io.Reader, out io.Writer) error {
+func (c *LockCommand) Exec(_ io.Reader, out io.Writer) error {
 	serviceID, serviceVersion, err := argparser.ServiceDetails(argparser.ServiceDetailsOpts{
-		Active:             optional.Of(false),
-		AutoCloneFlag:      c.autoClone,
+		Locked:             optional.Of(false),
 		APIClient:          c.Globals.APIClient,
 		Manifest:           *c.Globals.Manifest,
 		Out:                out,
@@ -76,15 +70,15 @@ func (c *ActivateCommand) Exec(_ io.Reader, out io.Writer) error {
 	c.Input.ServiceID = serviceID
 	c.Input.ServiceVersion = fastly.ToValue(serviceVersion.Number)
 
-	ver, err := c.Globals.APIClient.ActivateVersion(context.TODO(), &c.Input)
+	ver, err := c.Globals.APIClient.LockVersion(context.TODO(), &c.Input)
 	if err != nil {
 		c.Globals.ErrLog.AddWithContext(err, map[string]any{
 			"Service ID":      serviceID,
-			"Service Version": serviceVersion.Number,
+			"Service Version": fastly.ToValue(serviceVersion.Number),
 		})
 		return err
 	}
 
-	text.Success(out, "Activated service %s version %d", fastly.ToValue(ver.ServiceID), c.Input.ServiceVersion)
+	text.Success(out, "Locked service %s version %d", fastly.ToValue(ver.ServiceID), c.Input.ServiceVersion)
 	return nil
 }